Make creatematerial in one file
