--- python/Scripts/creatematerial/__init__.py
@@ -1,3 +1,0 @@
-
-
-
